Reorganize the layout of the code
