+++ package.json
@@ -2,8 +2,8 @@
   "name": "gensequence",
   "version": "2.1.3",
   "description": "Small library to simplify working with Generators and Iterators in Javascript / Typescript",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
   "dependencies": {},
   "devDependencies": {
     "@types/jest": "^24.0.23",
@@ -17,9 +17,9 @@
   "scripts": {
     "prepublish": "npm run clean-build && npm test",
     "clean-build": "npm run clean && npm run build",
-    "clean": "rimraf lib",
-    "test": "jest src/*.test.ts",
-    "perf": "jest src/*.perf.ts",
+    "clean": "rimraf dist",
+    "test": "jest src/**/*.test.ts",
+    "perf": "jest src/**/*.perf.ts",
     "build": "tsc -p .",
     "watch": "tsc -w -p .",
     "coverage": "npm test -- --coverage",
@@ -41,9 +41,9 @@
     "reduce"
   ],
   "files": [
-    "lib/",
-    "src/",
+    "dist/",
     "!**/samples",
+    "!**/*.map",
     "!**/*.test.*",
     "!**/*.perf.*"
   ],
