+++ src/GenSequence.ts
@@ -1,9 +1,5 @@
-
-export type Maybe<T> = T | undefined;
-
-export interface IterableLike<T> {
-    [Symbol.iterator](): Iterator<T> | IterableIterator<T>;
-}
+import { IterableLike, Maybe } from './operators/types';
+import { filter, skip, take, concat, concatMap, combine, map, scan, all, any, count, first, forEach, max, min, reduce } from './operators/operators';
 
 export interface Sequence<T> extends IterableLike<T> {
     next(): IteratorResult<T>;
@@ -57,13 +53,7 @@ export interface SequenceCreator<T> {
 export function genSequence<T>(i: () => GenIterable<T>): Sequence<T>;
 export function genSequence<T>(i: GenIterable<T>): Sequence<T>;
 export function genSequence<T>(i: (() => GenIterable<T>) | GenIterable<T>): Sequence<T> {
-    let createIterable: () => GenIterable<T>;
-    if (typeof i === "function") {
-        createIterable = i;
-    } else {
-        // this is typeof "object"
-        createIterable = () => i;
-    }
+    const createIterable: () => GenIterable<T> = (typeof i === "function") ? i : () => i;
 
     function fnNext() {
         let iter: Maybe<Iterator<T>>;
@@ -71,12 +61,7 @@ export function genSequence<T>(i: (() => GenIterable<T>) | GenIterable<T>): Sequ
             if(!iter) {
                 iter = createIterable()[Symbol.iterator]();
             }
-            const result: IteratorResult<T> = iter.next();
-            if (result.done) {
-                iter = undefined;
-            }
-
-            return result;
+            return iter.next();
         };
     }
 
@@ -145,7 +130,7 @@ export function genSequence<T>(i: (() => GenIterable<T>) | GenIterable<T>): Sequ
         //// Cast
         toArray: () => [...createIterable()],
         toIterable: () => {
-            return toIterator(createIterable());
+            return toIterableIterator(createIterable());
         },
     };
 
@@ -159,219 +144,15 @@ export const GenSequence = {
     sequenceFromObject,
 };
 
-//// Filters
-export function* filter<T>(fnFilter: (t: T) => boolean, i: Iterable<T>) {
-    for (const v of i) {
-        if (fnFilter(v)) {
-            yield v;
-        }
-    }
-}
-
-export function* skip<T>(n: number, i: Iterable<T>): IterableIterator<T> {
-    let a = 0;
-    for (const t of i) {
-        if (a >= n) {
-            yield t;
-        }
-        a += 1;
-    }
-}
-
-
-export function* take<T>(n: number, i: Iterable<T>): IterableIterator<T> {
-    let a = 0;
-    if (n) {
-        for (const t of i) {
-            if (a >= n) {
-                break;
-            }
-            yield t;
-            a += 1;
-        }
-    }
-}
-
-//// Extenders
-/**
- * Concat two iterables together
- */
-export function* concat<T>(i: Iterable<T>, j: Iterable<T>): IterableIterator<T> {
-    yield *i;
-    yield *j;
-}
-
-export function* concatMap<T, U>(fn: (t: T) => Iterable<U>, i: Iterable<T>): IterableIterator<U> {
-    for (const t of i) {
-        yield *fn(t);
-    }
-}
-
-//// Mappers
-/**
- * Combine two iterables together using fnMap function.
- */
-export function* combine<T, U, V>(fnMap: (t: T, u?: U) => V, i: Iterable<T>, j: Iterable<U>): IterableIterator<V> {
-    const jit = j[Symbol.iterator]();
-    for (const r of i) {
-        const s = jit.next().value;
-        yield fnMap(r, s);
-    }
-}
-
-/**
- * apply a mapping function to an Iterable.
- */
-export function map<T, U>(fnMap: (t: T) => U): (i: Iterable<T>) => IterableIterator<U>;
-export function map<T, U>(fnMap: (t: T) => U, i: Iterable<T>): IterableIterator<U>;
-export function map<T, U>(fnMap: (t: T) => U, i?: Iterable<T>): IterableIterator<U> | ((i: Iterable<T>) => IterableIterator<U>) {
-    function* fn<T, U>(fnMap: (t: T) => U, i: Iterable<T>): IterableIterator<U> {
-        for (const v of i) {
-            yield fnMap(v);
-        }
-    }
-
-    if (i !== undefined) {
-        return fn(fnMap, i);
-    }
-
-    return function(i: Iterable<T>) {
-        return fn(fnMap, i);
-    };
-}
-
-export function scan<T>(i: Iterable<T>, fnReduce: (prevValue: T, curValue: T, curIndex: number) => T): IterableIterator<T>;
-export function scan<T>(i: Iterable<T>, fnReduce: (prevValue: T, curValue: T, curIndex: number) => T, initValue: T): IterableIterator<T>;
-export function scan<T, U>(i: Iterable<T>, fnReduce: (prevValue: U, curValue: T, curIndex: number) => U, initValue: U): IterableIterator<U>;
-export function* scan<T>(i: Iterable<T>, fnReduce: (prevValue: T, curValue: T, curIndex: number) => T, initValue?: T): IterableIterator<T> {
-    let index = 0;
-    if (initValue === undefined) {
-        // We need to create a new iterable to prevent for...of from restarting an array.
-        index = 1;
-        const iter = i[Symbol.iterator]();
-        let r = iter.next();
-        if (!r.done) yield r.value;
-        initValue = r.value;
-        i = makeIterable(iter);
-    }
-    let prevValue: T = initValue!;
-    for (const t of i) {
-        const nextValue = fnReduce(prevValue, t, index);
-        yield nextValue;
-        prevValue = nextValue;
-        index += 1;
-    }
-}
-
-//// Reducers
-export function all<T>(fn: (t: T) => boolean, i: Iterable<T>): boolean {
-    for (const t of i) {
-        if (!fn(t)) {
-            return false;
-        }
-    }
-    return true;
-}
-
-export function any<T>(fn: (t: T) => boolean, i: Iterable<T>): boolean {
-    for (const t of i) {
-        if (fn(t)) {
-            return true;
-        }
-    }
-    return false;
-}
-
-export function count<T>(i: Iterable<T>): number {
-    return reduce<T, number>(p => p + 1, 0, i);
-}
-
-export function first<T>(fn: (t: T) => boolean, defaultValue: T, i: Iterable<T>): T;
-export function first<T>(fn: Maybe<(t: T) => boolean>, defaultValue: Maybe<T>, i: Iterable<T>): Maybe<T> {
-    fn = fn || (() => true);
-    for (const t of i) {
-        if (fn(t)) {
-            return t;
-        }
-    }
-    return defaultValue;
-}
-
-export function forEach<T>(fn: (t: T, index: number) => void, i: Iterable<T>) {
-    let index = 0;
-    for (const t of i) {
-        fn(t, index);
-        index += 1;
-    }
-}
-
-export function max<T, U>(selector: (t: T) => U, i: Iterable<T>): Maybe<T>;
-export function max<T>(selector: (t: T) => T = (t => t), i: Iterable<T>): Maybe<T> {
-    return reduce((p: T, c: T) => selector(c) > selector(p) ? c : p, undefined, i);
-}
-
-export function min<T, U>(selector: (t: T) => U, i: Iterable<T>): Maybe<T>;
-export function min<T>(selector: (t: T) => T = (t => t), i: Iterable<T>): Maybe<T> {
-    return reduce((p: T, c: T) => selector(c) < selector(p) ? c : p, undefined, i);
-}
-
-export function reduce<T, U>(fnReduce: (prevValue: U, curValue: T, curIndex: number) => U, initialValue: U, i: Iterable<T>): U;
-export function reduce<T>(fnReduce: (prevValue: T, curValue: T, curIndex: number) => T, initialValue: T, i: Iterable<T>): T;
-export function reduce<T>(fnReduce: (prevValue: T, curValue: T, curIndex: number) => T, initialValue: Maybe<T>, i: Iterable<T>): Maybe<T>;
-export function reduce<T>(fnReduce: (prevValue: T, curValue: T, curIndex: number) => T, initialValue: Maybe<T>, i: Iterable<T>): Maybe<T> {
-    let index = 0;
-    if (initialValue === undefined) {
-        index = 1;
-        const r = i[Symbol.iterator]().next();
-        initialValue = r.value;
-    }
-    let prevValue: T = initialValue!;
-    for (const t of i) {
-        const nextValue = fnReduce(prevValue, t, index);
-        prevValue = nextValue;
-        index += 1;
-    }
-    return prevValue;
-}
-
 //// Cast
-export function* toIterator<T>(i: Iterable<T>) {
+export function* toIterableIterator<T>(i: Iterable<T>) {
     yield* i;
 }
 
-//// Utilities
 /**
- * Convert an Iterator into an IterableIterator
+ * alias of toIterableIterator
  */
-export function makeIterable<T>(i: Iterator<T>) {
-    function* iterate() {
-        for (let r = i.next(); ! r.done; r = i.next()) {
-            yield r.value;
-        }
-    }
-    return iterate();
-}
-
-/**
- * Creates a scan function that can be used in a map function.
- */
-export function scanMap<T>(accFn: (acc: T, value: T) => T, init?: T): ((value: T) => T);
-export function scanMap<T, U>(accFn: (acc: U, value: T) => U, init: U): ((value: T) => U);
-export function scanMap<T>(accFn: (acc: T, value: T) => T, init?: T): ((value: T) => T) {
-    let acc = init;
-    let first = true;
-    return function(value: T): T {
-        if (first && acc === undefined) {
-            first = false;
-            acc = value;
-            return acc;
-        }
-        acc = accFn(acc as T, value);
-        return acc;
-    };
-}
-
-
+export const toIterator = toIterableIterator;
 
 export type KeyValuePair<T> = [keyof T, T[keyof T]];
