+++ src/index.ts
@@ -1,0 +1,2 @@
+export * from './GenSequence';
+export * from './operators/operators';
