+++ src/operators/types.ts
@@ -1,0 +1,5 @@
+export type Maybe<T> = T | undefined;
+
+export interface IterableLike<T> {
+    [Symbol.iterator](): Iterator<T> | IterableIterator<T>;
+}
