+++ src/operators/util.ts
@@ -1,0 +1,13 @@
+export function toIterator<T>(values: Iterable<T> | IterableIterator<T>) {
+    let iter: Iterator<T> | IterableIterator<T> | undefined;
+
+    const rangeIterator = {
+       next: function() {
+           if (!iter) {
+               iter = values[Symbol.iterator]();
+           }
+           return iter.next();
+       }
+    };
+    return rangeIterator;
+}
